Move the middleware in a separate class library project
